Move FebBase and FebEnum to feb_base.py
