+++ fepyio/mesh_domains.py
@@ -1,6 +1,6 @@
 from dataclasses import dataclass
 
-from ._base import FebBase
+from .feb_base import FebBase
 
 
 @dataclass
